+++ sdk/agentserver/Azure.AI.AgentServer.Invocations/api/Azure.AI.AgentServer.Invocations.net10.0.cs
@@ -17,9 +17,9 @@ namespace Azure.AI.AgentServer.Invocations
         public virtual System.Threading.Tasks.Task GetOpenApiAsync(Microsoft.AspNetCore.Http.HttpRequest request, Microsoft.AspNetCore.Http.HttpResponse response, System.Threading.CancellationToken cancellationToken) { throw null; }
         public abstract System.Threading.Tasks.Task HandleAsync(Microsoft.AspNetCore.Http.HttpRequest request, Microsoft.AspNetCore.Http.HttpResponse response, Azure.AI.AgentServer.Invocations.InvocationContext context, System.Threading.CancellationToken cancellationToken);
     }
-    public abstract partial class InvocationsWebSocketHandler : Azure.AI.AgentServer.Invocations.InvocationHandler
+    public abstract partial class InvocationWebSocketHandler : Azure.AI.AgentServer.Invocations.InvocationHandler
     {
-        protected InvocationsWebSocketHandler() { }
+        protected InvocationWebSocketHandler() { }
         public override System.Threading.Tasks.Task HandleAsync(Microsoft.AspNetCore.Http.HttpRequest request, Microsoft.AspNetCore.Http.HttpResponse response, Azure.AI.AgentServer.Invocations.InvocationContext context, System.Threading.CancellationToken cancellationToken) { throw null; }
         public abstract System.Threading.Tasks.Task HandleWebSocketAsync(System.Net.WebSockets.WebSocket webSocket, Azure.AI.AgentServer.Invocations.InvocationContext context, System.Threading.CancellationToken cancellationToken);
     }